dio merda

+++ gamifiedmarketing/src/main/java/it/polimi/db2/gamifiedmarketing/application/service/AdminService.java
@@ -1,12 +1,17 @@
 package it.polimi.db2.gamifiedmarketing.application.service;
 
 import it.polimi.db2.gamifiedmarketing.application.entity.Product;
+import it.polimi.db2.gamifiedmarketing.application.entity.Question;
 import it.polimi.db2.gamifiedmarketing.application.entity.Submission;
 import it.polimi.db2.gamifiedmarketing.application.entity.User;
+import it.polimi.db2.gamifiedmarketing.application.entity.enums.UserRole;
+import it.polimi.db2.gamifiedmarketing.application.entity.views.AddProductRequest;
 import it.polimi.db2.gamifiedmarketing.application.entity.views.ViewResponse;
 import it.polimi.db2.gamifiedmarketing.application.repository.ProductRepository;
 import it.polimi.db2.gamifiedmarketing.application.repository.SubmissionRepository;
 import it.polimi.db2.gamifiedmarketing.application.repository.UserRepository;
+import it.polimi.db2.gamifiedmarketing.application.session.SessionInfo;
+import it.polimi.db2.gamifiedmarketing.application.utility.Utils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -27,8 +32,19 @@ public class AdminService {
     @Autowired
     private ProductRepository productRepository;
 
+    @Autowired
+    private SessionInfo sessionInfo;
+
     public ViewResponse deleteProductByDate(LocalDate date) {
         try{
+            if (sessionInfo.getCurrentUser() == null) {
+                throw new Exception("You seems to not be logged in!");
+            }
+
+            if (!sessionInfo.getCurrentUser().getRole().equals(UserRole.ADMIN)) {
+                throw new Exception("You cannot delete a product if you are not admin");
+            }
+
             productRepository.deleteByDate(date);
             return new ViewResponse(true, null);
         }catch(Exception e){
@@ -38,9 +54,24 @@ public class AdminService {
         }
     }
 
-    public ViewResponse<Product> addProduct(Product product) {
+    public ViewResponse<Product> addProduct(AddProductRequest addProductRequest) {
         try{
-            Optional<User> user = userRepository.findById(product.getAdmin().getId());
+            if (sessionInfo.getCurrentUser() == null) {
+                throw new Exception("You seems to not be logged in!");
+            }
+
+            if (!sessionInfo.getCurrentUser().getRole().equals(UserRole.ADMIN)) {
+                throw new Exception("You cannot add a product if you are not admin");
+            }
+
+            Optional<User> user = userRepository.findById(sessionInfo.getCurrentUser().getId());
+            List<Question> questions = new ArrayList<Question>();
+            for(int i=0; i<addProductRequest.questions.size(); i++) {
+                Question question = Question.builder().title(addProductRequest.questions.get(i).new_question_title).subtitle(addProductRequest.questions.get(i).new_question_subtitle).build();
+                questions.add(question);
+            }
+
+            Product product = Product.builder().name(addProductRequest.name).description(addProductRequest.description).imageUrl(addProductRequest.image_url).questions(questions).build();
             product.setAdmin(user.get());
             Product _return = productRepository.save(product);
             return new ViewResponse(true, _return, null);
@@ -53,6 +84,14 @@ public class AdminService {
 
     public ViewResponse<List<Submission>> getVisualQuestionnaire(LocalDate date) {
         try{
+            if (sessionInfo.getCurrentUser() == null) {
+                throw new Exception("You seems to not be logged in!");
+            }
+
+            if (!sessionInfo.getCurrentUser().getRole().equals(UserRole.ADMIN)) {
+                throw new Exception("You cannot see all questionnaires if you are not admin");
+            }
+
             List<Submission>  _return = submissionRepository.getAllSubmissionOfTheDay(date);
             return new ViewResponse(true, _return, null);
         }catch(Exception e){
